chore: Add cache package and tests
This commit adds a new `cache` package along with its corresponding tests. The `cache` package includes a `Cache` struct that stores the results of functions, as well as methods to get and set values in the cache. Additionally, the commit introduces a `HashArgs` function that computes the MD5 hash of input arguments, and a `hashDirectory` function that computes the MD5 hash of all files in a directory.